+++ src/pages/Login/Login.tsx
@@ -24,7 +24,6 @@ const LoginForm = () => {
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    console.log("Form submitted");
 
     axios
       .post(