+++ src/redux/apis/alert.ts
@@ -14,8 +14,35 @@ const api = createApi({
         credentials: "include",
       }),
     }),
+    getAlerts: builder.query<AlertType[], void>({
+      query: () => ({
+        url: "/",
+        method: "GET",
+        credentials: "include",
+      }),
+    }),
+    deleteAlert: builder.mutation<void, number>({
+      query: (id) => ({
+        url: `/${id}/`,
+        method: "DELETE",
+        credentials: "include",
+      }),
+    }),
+    updateAlert: builder.mutation<AlertType, AlertType>({
+      query: (body) => ({
+        url: `/`,
+        method: "PUT",
+        body,
+        credentials: "include",
+      }),
+    }),
   }),
 });
 
-export const { useCreateAlertMutation } = api;
+export const {
+  useCreateAlertMutation,
+  useLazyGetAlertsQuery,
+  useDeleteAlertMutation,
+  useUpdateAlertMutation,
+} = api;
 export default api;